requirements.txt
discord.py==2.4.0
pytz==2024.2
python-dotenv==1.0.1